agritech-api/requirements.txt
cffi==1.14.2
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
Werkzeug==1.0.1
